sync

+++ plugins/weibo/model.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from sqlalchemy import ForeignKey, Integer
+from sqlalchemy import BigInteger, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base, Group
@@ -13,5 +13,5 @@ class WeiboFollower(Base):
     id: Mapped[str] = mapped_column(ForeignKey(Group.id, ondelete="CASCADE"), primary_key=True)
     """群组 ID"""
 
-    wid: Mapped[int] = mapped_column(Integer, primary_key=True)
+    wid: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     """微博用户 ID"""
